Version 1.0.0: first stable release

+++ build.sh
@@ -1,30 +1,42 @@
+# this build script is quite forceful about setup - make sure not to mess up the system python
+PYTHON_VENV=$(python -c "import sys; sys.stdout.write('1') if hasattr(sys, 'real_prefix') or sys.base_prefix != sys.prefix else sys.stdout.write('0')")
+if [ "$PYTHON_VENV" == 0 ]; then
+  echo 'Warning: not running in a Python virtual environment. Please either activate a venv or edit the script to confirm this action'
+  exit 1
+fi
+
 module=${PWD##*/} # get module name - relies on directory name == module name
 module=${module:-/}
 
-echo "Preparing environment for $module"
-python -m pip install twine # we don't have a requirements.txt so easiest just to be sure twine exists every time
+printf "\nPreparing environment for %s…\n" "$module"
+python -m pip install --quiet --upgrade pip
+python -m pip install --quiet setuptools wheel twine mypy sphinx # we don't have a requirements.txt, so easiest just to be sure build dependencies exist every time
 rm -rf dist
 rm -rf build
 
+python -m pip install --quiet --force-reinstall . # install pyoslog itself
+
 printf '\nType checking %s…\n' "$module"
-mypy pyoslog
+python -m mypy pyoslog
 
 printf '\nRunning tests for %s…\n' "$module"
-(cd tests && python -m unittest) # we don't fail on failed tests because they need a macOS version above our minimum
+python -m pip install --quiet pyobjc-framework-OSLog # separated from other installations because it will fail on unsupported platforms
+(cd tests && python -m unittest)                     # we don't fail on failed tests because they need a macOS version above our minimum
 
 printf '\nBuilding documentation for %s…\n' "$module"
-python -m pip install -r docs/requirements.txt
+python -m pip install --quiet -r docs/requirements.txt
+export SPHINXOPTS=-q
 (cd docs && make clean html)
 
-printf '\nBuilding source and wheel (universal) distributions for %s…\n' "$module"
-python setup.py clean --all sdist bdist_wheel --universal
+printf '\nBuilding source and wheel (universal) distributions for %s ("can'\''t clean" messages can be ignored)…\n' "$module"
+python setup.py -q clean --all sdist bdist_wheel --universal
 
 printf '\nValidating %s packages…\n' "$module"
-twine check dist/*
+python -m twine check dist/*
 
 if [ -z "$1" ]; then # exit unless a parameter is provided (we use 'deploy' but don't actually care what it is)
-  printf '\nExiting build script - run "./build.sh deploy" to also upload to PyPi / Read the Docs\n'
-  exit 1
+  printf '\nExiting build script - run "./build.sh deploy" to also upload to PyPi / Read the Docs (please git commit first)\n\n'
+  exit 0
 fi
 
 printf "\nUpload the $(tput bold)%s$(tput sgr0) package to the \033[0;32m$(tput bold)Test PyPI repository$(tput sgr0)\033[0m via Twine? (y to confirm; n to skip; any other key to exit): " "$module"
@@ -32,7 +44,7 @@ read -r answer
 
 if [ "$answer" != "${answer#[Yy]}" ]; then
   PYPI_TEST_TOKEN=$(<versions/pypi-test.token)
-  if twine upload -u __token__ -p "$PYPI_TEST_TOKEN" --repository testpypi dist/*; then
+  if python -m twine upload -u __token__ -p "$PYPI_TEST_TOKEN" --repository testpypi dist/*; then
     echo "Upload of $module completed – install via: python -m pip install --force-reinstall --index-url https://test.pypi.org/simple/ $module"
   else
     echo "Error uploading $module; exiting"
@@ -50,7 +62,7 @@ read -r answer
 
 if [ "$answer" != "${answer#[Yy]}" ]; then
   PYPI_DEPLOY_TOKEN=$(<versions/pypi-deploy.token)
-  if twine upload -u __token__ -p "$PYPI_DEPLOY_TOKEN" dist/*; then
+  if python -m twine upload -u __token__ -p "$PYPI_DEPLOY_TOKEN" dist/*; then
     echo "Upload of $module completed – view at https://pypi.org/project/$module"
 
     READTHEDOCS_TOKEN=$(<versions/readthedocs.token)
